qcacld-2.0: Correct/Increase len size of oem data request
oem data request len size should be of 4 byte instead of
1 byte. Add changes to correct len size in oem data structure across the
layers.

Increase maximum size of oem data request from 280 to 500. This
change set also remove the following unused macros in UMAC
- OEM_DATA_REQ_SIZE
- OEM_DATA_RSP_SIZE

Bug: 37948902

+++ drivers/staging/qcacld-2.0/CORE/WDA/inc/legacy/halMsgApi.h
@@ -752,7 +752,7 @@ typedef struct
 {
     tSirMacAddr          selfMacAddr;
     eHalStatus           status;
-    uint8_t              data_len;
+    uint32_t              data_len;
     uint8_t              *data;
 } tStartOemDataReq, *tpStartOemDataReq;
